requirements.txt
Pillow
uwsgi
Image
Flask
configparser
requests